chore: update package names and rename destination files
